requirements.txt
psycopg2-binary==2.8.4
PySocks==1.7.1
pyTelegramBotAPI==3.6.7